chore: cleanup api test code in kernel
