+++ package.json
@@ -113,7 +113,7 @@
     "@typescript-eslint/eslint-plugin": "5.51.0",
     "@typescript-eslint/parser": "5.51.0",
     "chalk": "4.1.2",
-    "chromedriver": "^111.0.0",
+    "chromedriver": "113.0.0",
     "codecov": "^3.8.3",
     "concurrently": "7.6.0",
     "cross-env": "7.0.3",
